+++ scout/commands/export/variant.py
@@ -1,7 +1,10 @@
 import click
 import logging
 
-from scout.export.variant import export_causatives
+from bson.json_util import dumps
+
+from scout.export.variant import export_variants
+from .utils import json_option
 
 LOG = logging.getLogger(__name__)
 
@@ -9,26 +12,35 @@ LOG = logging.getLogger(__name__)
 @click.option('-c', '--collaborator',
         help="Specify what collaborator to export variants from. Defaults to cust000",
 )
+@click.option('-d', '--document-id',
+        help="Search for a specific variant",
+)
+@json_option
 @click.pass_context
-def variants(context, collaborator):
+def variants(context, collaborator, document_id, json):
     """Export causatives for a collaborator in .vcf format"""
     LOG.info("Running scout export variants")
     adapter = context.obj['adapter']
     collaborator = collaborator or 'cust000'
-
-    header = [
+    
+    variants = export_variants(adapter, collaborator, document_id=document_id)
+    
+    if json:
+        click.echo(dumps([var for var in variants]))
+        return
+    
+    vcf_header = [
         "##fileformat=VCFv4.2",
         "#CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO"
     ]
 
-    for line in header:
+    for line in vcf_header:
         click.echo(line)
 
     #put variants in a dict to get unique ones
     variant_string = ("{0}\t{1}\t{2}\t{3}\t{4}\t{5}\t{6}\t{7}")
 
-    ##TODO add option to give output in json format
-    for variant_obj in export_causatives(adapter, collaborator):
+    for variant_obj in variants:
         click.echo(variant_string.format(
             variant_obj['chromosome'],
             variant_obj['position'],